# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  common.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  dto.CreateMemberLevelRequest:
    properties:
      discount:
        maximum: 10
        minimum: 1
        type: number
      icon:
        maxLength: 256
        type: string
      minPoints:
        type: integer
      name:
        maxLength: 64
        type: string
      sort:
        type: integer
      status:
        enum:
        - 0
        - 1
        type: integer
    required:
    - name
    type: object
...
  dto.CreateMemberTagRequest:
    properties:
      color:
        maxLength: 20
        type: string
      name:
        maxLength: 64
        type: string
      sort:
        type: integer
      status:
        enum:
        - 0
        - 1
        type: integer
    required:
    - name
    type: object
...
  dto.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  dto.RefreshTokenRequest:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
...
  dto.UpdateMemberLevelRequest:
    properties:
      discount:
        maximum: 10
        minimum: 1
        type: number
      icon:
        maxLength: 256
        type: string
      id:
        type: integer
      minPoints:
        type: integer
      name:
        maxLength: 64
        type: string
      sort:
        type: integer
      status:
        enum:
        - 0
        - 1
        type: integer
    required:
    - id
    - name
    type: object
...
  dto.UpdateMemberTagRequest:
    properties:
      color:
        maxLength: 20
        type: string
      id:
        type: integer
      name:
        maxLength: 64
        type: string
      sort:
        type: integer
      status:
        enum:
        - 0
        - 1
        type: integer
    required:
    - id
    - name
    type: object
...
  vo.LoginResponse:
    properties:
      accessToken:
        type: string
      expiresIn:
        type: integer
      refreshToken:
        type: string
      tokenType:
        type: string
    type: object
  vo.MenuInfo:
    properties:
      children:
        items:
          $ref: '#/definitions/vo.MenuInfo'
        type: array
      component:
        type: string
      icon:
        type: string
      id:
        type: integer
      isCache:
        type: integer
      name:
        type: string
      parentId:
        type: integer
      path:
        type: string
      redirect:
        type: string
      sort:
        type: integer
      title:
        type: string
      type:
        type: integer
      visible:
        type: integer
    type: object
  vo.RoleInfo:
    properties:
      code:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  vo.UserInfoResponse:
    properties:
      avatar:
        type: string
      buttons:
        items:
          type: string
        type: array
      email:
        type: string
      id:
        type: integer
      menus:
        items:
          $ref: '#/definitions/vo.MenuInfo'
        type: array
      nickname:
        type: string
      phone:
        type: string
      roles:
        items:
          $ref: '#/definitions/vo.RoleInfo'
        type: array
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Gin-Admin 后台管理系统 API 文档
  title: Gin-Admin API
  version: 1.0.0
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 登录参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/vo.LoginResponse'
              type: object
      summary: 用户登录
      tags:
      - 认证
  /api/v1/auth/logout:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
      summary: 退出登录
      tags:
      - 认证
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: RefreshToken
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/vo.LoginResponse'
              type: object
      summary: 刷新Token
      tags:
      - 认证
  /api/v1/auth/userInfo:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/vo.UserInfoResponse'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取用户信息
      tags:
      - 认证
...
  /member/level:
    post:
      consumes:
      - application/json
      parameters:
      - description: 等级信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateMemberLevelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 创建会员等级
      tags:
      - 会员等级
    put:
      consumes:
      - application/json
      parameters:
      - description: 等级信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateMemberLevelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 更新会员等级
      tags:
      - 会员等级
  /member/level/{id}:
    delete:
      parameters:
      - description: 等级ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 删除会员等级
      tags:
      - 会员等级
...
  /member/level/list:
    get:
      parameters:
      - description: 等级名称
        in: query
        name: name
        type: string
      - description: 页码
        in: query
        name: page
        required: true
        type: integer
      - description: 每页条数
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 等级列表
      tags:
      - 会员等级
...
  /member/points/list:
    get:
      parameters:
      - description: 会员ID
        in: query
        name: memberId
        type: integer
      - description: 类型 1获取 2消费
        in: query
        name: type
        type: integer
      - description: 页码
        in: query
        name: page
        required: true
        type: integer
      - description: 每页条数
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 积分明细列表
      tags:
      - 积分明细
...
  /member/tag:
    post:
      consumes:
      - application/json
      parameters:
      - description: 标签信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateMemberTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 创建会员标签
      tags:
      - 会员标签
    put:
      consumes:
      - application/json
      parameters:
      - description: 标签信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateMemberTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 更新会员标签
      tags:
      - 会员标签
  /member/tag/{id}:
    delete:
      parameters:
      - description: 标签ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 删除会员标签
      tags:
      - 会员标签
...
  /member/tag/list:
    get:
      parameters:
      - description: 标签名称
        in: query
        name: name
        type: string
      - description: 页码
        in: query
        name: page
        required: true
        type: integer
      - description: 每页条数
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 标签列表
      tags:
      - 会员标签
...
  /system/file/{id}:
    delete:
      parameters:
      - description: 文件ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 删除文件
      tags:
      - 文件
    get:
      parameters:
      - description: 文件ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 获取文件详情
      tags:
      - 文件
  /system/file/list:
    get:
      parameters:
      - description: 文件名
        in: query
        name: name
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页条数
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 文件列表
      tags:
      - 文件
  /system/file/upload:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 上传文件
      tags:
      - 文件
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: '输入格式: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"